Added Mutant class, fixed HP buttons
Added the Mutant class, and fixed the HP buttons by adding a new button for Heal along with the one for Level - the players were using the Level button to roll their hit points when they got healed causing them to be healed for more (or less) depending on Stamina bonuses and Lucky Rolls'.  Those bonuses are not included in regular HP rolls.